+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 2.1.1 (TBD)
 
-* TBD
+* Point to MGRS conversion latitude min/max bounding and longitude normalization
 
 ## [2.1.0](https://github.com/ngageoint/mgrs-java/releases/tag/2.1.0) (09-22-2022)